backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
networkx>=3.2
scikit-learn>=1.4.0
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
anthropic>=0.25.0
python-dotenv>=1.0.0
